Renaming indicator_tags file into tags.py
