+++ trustar2/handlers/__init__.py
@@ -1,4 +1,4 @@
-from trustar2.handlers.indicator_tags import TagIndicator
+from trustar2.handlers.tags import TagIndicator
 from trustar2.handlers.indicators import SearchIndicator
 from trustar2.handlers.safelist import Safelist
 from trustar2.handlers.submission import Submission
