+++ trustar2/handlers/submission.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 from trustar2.query import Query
 from trustar2.trustar_enums import MaxValues
 from trustar2.handlers.base_handler import BaseHandler
+from trustar2.handlers.tags import TagSubmission
 from trustar2.base import fluent, Methods, get_timestamp
 
 
@@ -297,4 +298,4 @@ class Submission(BaseHandler):
 
 
     def tags(self):
-        return TagIndicator(self.config)
+        return TagSubmission(self.config)
